Fixed reporting of very large file sizes

+++ src/header_details_dialog.cpp
@@ -50,7 +50,7 @@ c_header_details_dialog::c_header_details_dialog(QWidget *parent)
 
 void c_header_details_dialog::set_details(
         QString filename,
-        int filesize,
+        int64_t filesize,
         QString file_id,
         int lu_id,
         int colour_id,
